Update to take advantage of the update to APNS2 which supports Auth Tokens, so multiple applications can share the same Authorization Token rather than creating a certificate for each application.
The API is unchanged, but to be more generic, the parameter to
the functions in apns.py was changed from certificate to creds.

The major change ended up being in the configuration, since FCM and
APNS went in the direction of tokens to simplify for multiple apps, and
the app configuration went in the direction of many certificates, one
for each app. So, the future API is going to get extremely cumbersome,
which will have to be addressed.