Update to take advantage of the update to APNS2 which supports Auth Tokens, so multiple applications can share the same Authorization Token rather than creating a certificate for each application.
The API is unchanged, but to be more generic, the parameter to
the functions in apns.py was changed from certificate to creds.

The major change ended up being in the configuration, since FCM and
APNS went in the direction of tokens to simplify for multiple apps, and
the app configuration went in the direction of many certificates, one
for each app. So, the future API is going to get extremely cumbersome,
which will have to be addressed.

Fixed formatting caught by flake8.

Second try at addressing flake8 requirements.

 o Updated to change variables to be PEP8 happy.
 o removed get_apns_token_lifetime. We can always expose it later.
 o Removed some overly-aspirational comments
 o Allow failures for djangomaster (2.11) since DRF doesn’t support it
   yet.

Missing a line continuation.