registration_id no longer hard-coded in DeviceViewSetMixin
`lookup_field` was declared in `DeviceViewSetMixin` and never used. Instead the create method referred to string literal instead of this field. No the field is used. You can inherit from `DeviceViewSetMixin`, redefined the `lookup_field` and use if with your views instead of having to rewrite the `create` method yourself