More robust handling of userAgentData, fix WebPushDeviceAdmin (#643)
* webpush add Edge

* webpush -  add edge endpoint

* change edge endpoint

* update README

Show another, easier approach to extract the browsername and browser version

* more robust handling ofh userAgentData

Relevant userAgentData ist not always stored in navigator.userAgentData.brands[0]

It is necessary to loop through all available navigator.userAgentData.brands and try to find the relevant entry. In case of Edge it is also necessary to ignore any matches regarding chrome...

addresses #640

* Fix WebpushDeviceAdmin

Field "device_id" is not available for WebPushDevices. Instead add "browser" to ListView and "registration_id" to search_fields