feat: Create module to support library-specific enumeration types
Motivation:

We never know when we might be implementing or extending the
`django-push-notifications` library beyond its current support.

It makes sense to have a module to handle critical moments such as
that raised in #708 (for  `interruption-levels`). This gives control
over newly  implemented interfaces without having to worry about
outdated modules, such as `PyAPNS2`.

This allows us to provide wrapper or extend existing library
functionalities.

Context:

This patch address one side of issue by providing an enumeration module
that can be used to deal with public and underlying APIs.