Fix #744: Bulk messages with aoiapns not working properly
Previous version implemented aioapns in a way that made
it hang indefinetly. Especially when receiver list contaned
a lot of bad tokens.

Having a lot of bad tokens still affects the reliability and
transfer speed of notification sets, which is why this fix
also deactivate devices for error codes BadDeviceToken and
DeviceTokenNotForTopic unlike previous versions.