feat(dependabot): automate dependency management and security updates
This PR is to address the re-occuring issue of managing dependencies or
transitive dependency for the project - including but not limited to,
security updates.

With dependabot, we are one step closer with keeping underlying
dependencies up-to-date, minimizing the overhead of manual intervention.

This would require constant review of the bot PRs, to ensure updates or security updates
are in-line with future release and backward compatible with this
package/libarary.